# outputs/requirement-clarify-manager/04-dev/backup_before/backend/src/main/kotlin/org/openapitools/api/ApiApiController.kt
package org.openapitools.api

import org.openapitools.model.AnalysisViewResponse
import org.openapitools.model.ApiV1ProjectsProjectNameDataDictionaryTermIdClarifyPost200Response
import org.openapitools.model.ApiV1RequirementsIdStatusPatch200Response
import org.openapitools.model.ApiV1RequirementsIdStatusPatchRequest
import org.openapitools.model.AsyncJobResponse
import org.openapitools.model.ClarifyRequest
import org.openapitools.model.InlineObject
import org.openapitools.model.PendingTermResponse
import org.openapitools.model.RequirementResponse
import io.swagger.v3.oas.annotations.*
import io.swagger.v3.oas.annotations.enums.*
import io.swagger.v3.oas.annotations.media.*
import io.swagger.v3.oas.annotations.responses.*
import io.swagger.v3.oas.annotations.security.*
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import jakarta.validation.Valid
import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size

import kotlin.collections.List
import kotlin.collections.Map

@RestController
@Validated
@RequestMapping("\${api.base-path:}")
class ApiApiController() {

    @Autowired
    lateinit var requirementService: org.openapitools.service.RequirementService

    @Operation(
        summary = "미정의 용어 단건 조회",
        operationId = "apiV1ProjectsProjectNameDataDictionaryPendingGet",
        description = """프로젝트 내 미정의(Undefine) 상태인 데이터 사전 용어 중 다음 명확화 대상 1건을 조회합니다.""",
        responses = [
            ApiResponse(responseCode = "200", description = "OK", content = [Content(schema = Schema(implementation = PendingTermResponse::class))]) ],
        security = [ SecurityRequirement(name = "BearerAuth") ]
    )
    @RequestMapping(
        method = [RequestMethod.GET],
        value = ["/api/v1/projects/{projectName}/data-dictionary/pending"],
        produces = ["application/json"]
    )
    fun apiV1ProjectsProjectNameDataDictionaryPendingGet(@Parameter(description = "", required = true) @PathVariable("projectName") projectName: kotlin.String): ResponseEntity<PendingTermResponse> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }

    @Operation(
        summary = "단일 용어 명확화 제출 (인터랙션 피드백)",
        operationId = "apiV1ProjectsProjectNameDataDictionaryTermIdClarifyPost",
        description = """기획자가 특정 용어에 대해 정의(DEFINE)를 수행하거나 건너뛰기(SKIP)를 선택합니다. 응답으로 즉시 다음 미정의 단어를 반환합니다.""",
        responses = [
            ApiResponse(responseCode = "200", description = "OK", content = [Content(schema = Schema(implementation = ApiV1ProjectsProjectNameDataDictionaryTermIdClarifyPost200Response::class))]),
            ApiResponse(responseCode = "404", description = "존재하지 않는 용어 ID"),
            ApiResponse(responseCode = "409", description = "동시성 충돌 (Optimistic Lock 예외)") ],
        security = [ SecurityRequirement(name = "BearerAuth") ]
    )
    @RequestMapping(
        method = [RequestMethod.POST],
        value = ["/api/v1/projects/{projectName}/data-dictionary/{termId}/clarify"],
        produces = ["application/json"],
        consumes = ["application/json"]
    )
    fun apiV1ProjectsProjectNameDataDictionaryTermIdClarifyPost(@Parameter(description = "", required = true) @PathVariable("projectName") projectName: kotlin.String,@Parameter(description = "", required = true) @PathVariable("termId") termId: kotlin.Int,@Parameter(description = "", required = true) @Valid @RequestBody clarifyRequest: ClarifyRequest): ResponseEntity<ApiV1ProjectsProjectNameDataDictionaryTermIdClarifyPost200Response> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }

    @Operation(
        summary = "통합 분석 결과 뷰어 조회",
        operationId = "apiV1RequirementsIdAnalysisViewGet",
        description = """""",
        responses = [
            ApiResponse(responseCode = "200", description = "OK", content = [Content(schema = Schema(implementation = AnalysisViewResponse::class))]) ],
        security = [ SecurityRequirement(name = "BearerAuth") ]
    )
    @RequestMapping(
        method = [RequestMethod.GET],
        value = ["/api/v1/requirements/{id}/analysis-view"],
        produces = ["application/json"]
    )
    fun apiV1RequirementsIdAnalysisViewGet(@Parameter(description = "", required = true) @PathVariable("id") id: kotlin.Int): ResponseEntity<AnalysisViewResponse> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }

    @Operation(
        summary = "요구사항 상세 Event Storming 분석 요청 (비동기)",
        operationId = "apiV1RequirementsIdAnalyzePost",
        description = """완성된 데이터 사전을 바탕으로 Event Storming 요소들을 추출합니다.""",
        responses = [
            ApiResponse(responseCode = "202", description = "Accepted (Processing)", content = [Content(schema = Schema(implementation = AsyncJobResponse::class))]),
            ApiResponse(responseCode = "422", description = "Unprocessable Entity (데이터 사전 미완료)") ],
        security = [ SecurityRequirement(name = "BearerAuth") ]
    )
    @RequestMapping(
        method = [RequestMethod.POST],
        value = ["/api/v1/requirements/{id}/analyze"],
        produces = ["application/json"]
    )
    fun apiV1RequirementsIdAnalyzePost(@Parameter(description = "", required = true) @PathVariable("id") id: kotlin.Int): ResponseEntity<AsyncJobResponse> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }

    @Operation(
        summary = "데이터 사전 자동 추출 요청 (비동기)",
        operationId = "apiV1RequirementsIdDataDictionaryExtractPost",
        description = """""",
        responses = [
            ApiResponse(responseCode = "202", description = "Accepted (Processing)", content = [Content(schema = Schema(implementation = AsyncJobResponse::class))]) ],
        security = [ SecurityRequirement(name = "BearerAuth") ]
    )
    @RequestMapping(
        method = [RequestMethod.POST],
        value = ["/api/v1/requirements/{id}/data-dictionary/extract"],
        produces = ["application/json"]
    )
    fun apiV1RequirementsIdDataDictionaryExtractPost(@Parameter(description = "", required = true) @PathVariable("id") id: kotlin.Int): ResponseEntity<AsyncJobResponse> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }

    @Operation(
        summary = "요구사항 상태 변경",
        operationId = "apiV1RequirementsIdStatusPatch",
        description = """""",
        responses = [
            ApiResponse(responseCode = "200", description = "OK", content = [Content(schema = Schema(implementation = ApiV1RequirementsIdStatusPatch200Response::class))]),
            ApiResponse(responseCode = "400", description = "Bad Request (유효성 검증 실패 등)", content = [Content(schema = Schema(implementation = InlineObject::class))]) ],
        security = [ SecurityRequirement(name = "BearerAuth") ]
    )
    @RequestMapping(
        method = [RequestMethod.PATCH],
        value = ["/api/v1/requirements/{id}/status"],
        produces = ["application/json"],
        consumes = ["application/json"]
    )
    fun apiV1RequirementsIdStatusPatch(@Parameter(description = "", required = true) @PathVariable("id") id: kotlin.Int,@Parameter(description = "", required = true) @Valid @RequestBody apiV1RequirementsIdStatusPatchRequest: ApiV1RequirementsIdStatusPatchRequest): ResponseEntity<ApiV1RequirementsIdStatusPatch200Response> {
        val result = requirementService.updateStatus(id, apiV1RequirementsIdStatusPatchRequest)
        return ResponseEntity.ok(result)
    }

    @Operation(
        summary = "요구사항 등록",
        operationId = "apiV1RequirementsPost",
        description = """마크다운 파일(.md)을 업로드하거나 텍스트를 직접 입력하여 요구사항을 등록합니다.""",
        responses = [
            ApiResponse(responseCode = "201", description = "Created", content = [Content(schema = Schema(implementation = RequirementResponse::class))]),
            ApiResponse(responseCode = "400", description = "Bad Request (유효성 검증 실패 등)", content = [Content(schema = Schema(implementation = InlineObject::class))]),
            ApiResponse(responseCode = "413", description = "Payload Too Large (10MB 초과)") ],
        security = [ SecurityRequirement(name = "BearerAuth") ]
    )
    @RequestMapping(
        method = [RequestMethod.POST],
        value = ["/api/v1/requirements"],
        produces = ["application/json"],
        consumes = ["multipart/form-data"]
    )
    fun apiV1RequirementsPost(@Parameter(description = "요구사항 메타데이터 및 상세 텍스트 (JSON 문자열)", required = true) @Valid @RequestParam(value = "data", required = true) `data`: kotlin.String ,@Parameter(description = "마크다운(.md) 파일 (최대 10MB)") @Valid @RequestPart("file", required = false) file: org.springframework.web.multipart.MultipartFile?): ResponseEntity<RequirementResponse> {
        val result = requirementService.createRequirement(data, file)
        return ResponseEntity.status(HttpStatus.CREATED).body(result)
    }
}