# outputs/requirement-clarify-manager/04-dev/backup_before/backend/src/main/kotlin/org/openapitools/model/AsyncJobResponse.kt
package org.openapitools.model

import java.util.Objects
import com.fasterxml.jackson.annotation.JsonProperty
import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size
import jakarta.validation.Valid
import io.swagger.v3.oas.annotations.media.Schema

/**
 * 
 * @param jobId 
 * @param status 
 * @param message 
 */
data class AsyncJobResponse(

    @Schema(example = "null", description = "")
    @get:JsonProperty("jobId") val jobId: kotlin.String? = null,

    @Schema(example = "null", description = "")
    @get:JsonProperty("status") val status: kotlin.String? = null,

    @Schema(example = "null", description = "")
    @get:JsonProperty("message") val message: kotlin.String? = null
    ) {

}